Game finished.
function checkGameOver() added to check if either player has reached 5 points. If so,
a message will display who won through the output id and an alert. Also, if there is a winner, the variable checkGameOver
will be set to true.

In playRound(), the function will check if gameOver is true, and reset all scores, changing
the gameOver variable back to false.